fix.customer 삭제 후 member에 통합
경로도 다 수정 후 노션에 반영했습니다.
