--- src/main/java/com/grepp/team07/app/model/customer/CustomerRepository.java
@@ -1,15 +1,0 @@
-package com.grepp.team07.app.model.customer;
-
-import com.grepp.team07.app.model.customer.dto.CustomerDto;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-@Mapper
-public interface CustomerRepository {
-    CustomerDto findById(@Param("userId") String userId);
-
-
-    void update(@Param("userId")String userId, @Param("dto")CustomerDto dto);
-
-    String findEncodedPasswordByUserId(String userId);
-}
