--- src/main/resources/mybatis/mappers/customer/customerMapper.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.grepp.team07.app.model.customer.CustomerRepository">
-
-    <resultMap id="customerMap" type="CustomerDto">
-        <id property="userId" column="user_id"/>
-    </resultMap>
-
-    <select id="findById" resultMap="customerMap">
-        select * from customer
-        where user_id = #{userId}
-    </select>
-
-    <update id="update">
-        update customer
-        set
-        email = #{dto.email},
-        password = #{dto.password},
-        address = #{dto.address},
-        post_code = #{dto.postCode}
-        where user_id = #{userId}
-    </update>
-    <update id="updateDeliveredAt" parameterType="DeliveryDto">
-        update delivery set delivered_at = #{deliveredAt}
-        where delivery_id = #{deliveryId}
-    </update>
-    <select id="findEncodedPasswordByUserId" resultType="string">
-        SELECT password FROM customer WHERE user_id = #{userId}
-    </select>
-</mapper>
